T1: Added a menu block in between trainer and wild battles in game-loop.js. Noticed that playerData did NOT need to be returned from mainMenu function and then currentPlayerData set to it - they were references to the same object, so making changes to currentPlayerData in menu automatically changes it in game-loop.js. Added a todo to clean up unnecessary implementations of this later.
Minor fix to setNextPokemon in battle.js so that it iterates through pokemonList instead of belt and only does so if the current pokeball has no health. This means that switching pokemon in menu.js which changes currentPokeball actually affects battle. Previously, it was iterating through and always setting currentPokeball to first ball in belt with health.